Rosary Ring
Rosary Ring is a simple Flutter app designed to simulate the experience of using a traditional rosary to count prayers or meditation cycles. The app features a circular counter and intuitive controls to increment and reset the count, all within a visually appealing interface.

Features
Interactive Counter: A circular counter that visually displays the current count in the center of the screen.
Increment and Reset Controls: Easy-to-use buttons to increment and reset the count to zero.
Responsive Design: The design adapts seamlessly to different screen sizes.
Arabic Quote Display: Displays a verse from the Holy Quran at the top of the screen.
Project Structure
main.dart: The main entry point to the app. Launches the app and sets up the home screen.
Home Screen: The home screen where the counter is displayed and managed.